请求steam增加ip池

+++ yudao-module-steam/yudao-module-steam-biz/src/main/java/cn/iocoder/yudao/module/steam/dal/dataobject/bindipaddress/BindIpaddressDO.java
@@ -30,7 +30,7 @@ public class BindIpaddressDO extends BaseDO {
     /**
      * 端口
      */
-    private Long port;
+    private Integer port;
     /**
      * 地址池id
      */
